pyproject.toml
[project]
name = "pydna"
dynamic = ["version"]
description = "Representing double stranded DNA and functions for simulating cloning and homologous recombination between DNA molecules."
authors = ["Björn F. Johansson<bjornjobb@gmail.com>"]
readme = "README.md"
classifiers=["Development Status :: 4 - Beta",
             "Environment :: Console",
             "Intended Audience :: Education",
             "Intended Audience :: Developers",
             "Intended Audience :: Science/Research",
             "License :: OSI Approved :: BSD License",
             "Operating System :: OS Independent",
             "Programming Language :: Python :: 3.7",
             "Programming Language :: Python :: 3.8",
             "Programming Language :: Python :: 3.9",
             "Programming Language :: Python :: 3.10",
             "Programming Language :: Python :: 3.11",
             "Topic :: Education",
             "Topic :: Scientific/Engineering :: Bio-Informatics", ]
[project.urls]
homepage = "https://github.com/BjornFJohansson/pydna#-pydna"
documentation = "https://pydna.readthedocs.io/?badge=latest"
repository = "https://github.com/BjornFJohansson/pydna/tree/dev_bjorn"
changelog = "https://github.com/BjornFJohansson/pydna/blob/master/docs/CHANGELOG.md#changelog"
[tool.poetry]
name = "pydna"
version = "5.2.0-a.15"
description = "Representing double stranded DNA and functions for simulating cloning and homologous recombination between DNA molecules."
authors = ["BjornFJohansson <bjornjobb@gmail.com>"]
license = "BSD"
readme = "README.md"
[tool.poetry.dependencies]
python = ">=3.8"
appdirs = ">=1.4.4"
biopython = ">=1.80"
networkx = ">=2.8.8"
pydivsufsort = ">=0.0.11"
prettytable = ">=3.5.0"
pyperclip = ">=1.8.2"
pyfiglet = ">=0.8.post1"
scipy = { version = ">=1.8.0", optional = true }
matplotlib = { version = ">=3.4.3", optional = true }
pillow = { version = ">=8.4.0", optional = true }
pyparsing = { version = ">=2.4.7", optional = true }
requests = { version = ">=2.26.0", optional = true }
cai2 = { version = ">=1.0.5", optional = true }
pyqt5 = { version = ">=5.15.0", optional = true }

[tool.poetry.extras]
gel = ["scipy", "matplotlib", "pillow"]
download = ["pyparsing", "requests"]
gui = ["pyqt5"]
express = ["cai2"]
[build-system]
requires = ["poetry-core",
            "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
[tool.poetry-dynamic-versioning.substitution]
folders = [
  { path = "src"}
]
[tool.poetry.group.test] # This part can be left out
[tool.poetry.group.test.dependencies]
pytest = ">=7.2.0"
pytest-cov = ">=4.0.0"
pytest-doctestplus = ">=0.12.1"
pytest-profiling = ">=1.7.0"
coverage = ">=7.1.0"
nbval = ">=0.9.6"
requests-mock = ">=1.10.0"
pyparsing = ">=2.4.7"
[tool.pytest.ini_options]
minversion = "6.0.2"
python_files = "test_*.py"
testpaths = [
    "tests",
    "src",
]